+++ src/qt/bitcoin.cpp
@@ -30,7 +30,7 @@
 #include "util.h"
 
 #ifdef ENABLE_WALLET
-#include "wallet.h"
+#include "wallet/wallet.h"
 #endif
 
 #include <stdint.h>
