+++ src/qt/coincontroldialog.cpp
@@ -15,7 +15,7 @@
 
 #include "coincontrol.h"
 #include "main.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 
 #include <boost/assign/list_of.hpp> // for 'map_list_of()'
