+++ src/qt/sendcoinsdialog.cpp
@@ -18,7 +18,7 @@
 #include "base58.h"
 #include "coincontrol.h"
 #include "ui_interface.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 
 #include <QMessageBox>
 #include <QScrollBar>
