+++ src/qt/splashscreen.cpp
@@ -12,7 +12,7 @@
 #include "version.h"
 
 #ifdef ENABLE_WALLET
-#include "wallet.h"
+#include "wallet/wallet.h"
 #endif
 
 #include <QApplication>
