+++ src/test/rpc_wallet_tests.cpp
@@ -6,7 +6,7 @@
 #include "rpcclient.h"
 
 #include "base58.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 
 #include "test/test_bitcoin.h"
