+++ src/wallet/wallet.h
@@ -14,8 +14,8 @@
 #include "keystore.h"
 #include "main.h"
 #include "ui_interface.h"
-#include "wallet_ismine.h"
-#include "walletdb.h"
+#include "wallet/wallet_ismine.h"
+#include "wallet/walletdb.h"
 
 #include <algorithm>
 #include <map>
