+++ src/wallet/walletdb.h
@@ -7,7 +7,7 @@
 #define BITCOIN_WALLETDB_H
 
 #include "amount.h"
-#include "db.h"
+#include "wallet/db.h"
 #include "key.h"
 #include "keystore.h"
